deps: Support non-breaking versions of openfeature-sdk for meta provider
The meta provider's gemspec stated only v0.3.x could be used. Unfortunately, this
conflicts with the official LaunchDarkly provider, which uses version 0.4.0, preventing
this gem from being used along with it. This update allows either 0.3.0 or 0.4.0 to be used.

Although technically speaking openfeature-sdk 0.3.0 -> 0.4.0 is a breaking change
because of the change of string values. However, the provider uses only the constants rather than
the string values so this should not break the provider.

Signed-off-by: Jordan Blacker <jblacker@justworks.com>